MetMashR: Metabolite Mashing with R

A package to merge, filter sort, organise and otherwise mash together metabolite annotation tables. Metabolite annotations can be imported from multiple sources (software) and combined using workflow steps based on S4 class templates derived from the `struct` package. Other modular workflow steps such as filtering, merging, splitting, normalisation and rest-api queries are included.

See also

Author

Maintainer: Gavin Rhys Lloyd g.r.lloyd@bham.ac.uk (ORCID)

Authors:

Ralf Johannes Maria Weber r.j.weber@bham.ac.uk