An annotation database

An annotation_database is an annotation_source() where the imported data.frame contains meta data for annotations. For example it might be a table of molecular identifiers, associated pathways etc.

Usage

annotation_source(source = character(0), data = data.frame(), tag = "", ...)

Arguments

data

(data.frame, NULL) A data.frame of annotation data. The default is data.frame().

tag

(character) A (short) character string that is used to represent this source e.g. in column names or source columns when used in a workflow. The default is "".

...

Additional slots and values passed to struct_class.

Value

A annotation_database object. This object has no output slots.

Inheritance

A annotation_database object inherits the following struct classes:

[annotation_database] -> [annotation_source] -> [struct_class]

See also

Examples

M = annotation_database( tag = character(0), data =
data.frame(id=NA), source = "ANY")