Filter by factor levels

Removes (or includes) annotations such that the named column excludes (or includes) the specified levels.

Usage

filter_venn(
  factor_name,
  group_column = NULL,
  tables = NULL,
  levels,
  mode = "exclude",
  perl = FALSE,
  fixed = FALSE,
  ...
)

Arguments

factor_name

(character) The name of the column(s) in the annotation_source to generate a chart from. Up to seven columns can be compared for a single annotation_source.

group_column

(character, NULL) The name of the column in the annotation_source to create groups from in the Venn diagram. This parameter is ignored if !is.null(tables), as each table is considered to be a group. This parameter is also ignored if more than one factor_name is provided, as each column is considered a group. The default is NULL.

tables

(list, NULL) A list of annotation_sources to generate the venn groups from. If the only table of interest is the table coming in from model_apply then set tables = NULL and use group_column. The default is NULL.

levels

(character) The venn diagram levels to filter by.

mode

(character) Filter mode. Allowed values are limited to the following:

"exclude": The specified levels are removed from the annotation table.

"include": Only the specified levels are retained in the annotation table.

The default is "exclude".

perl

(logical) Use a Perl-compatible regex. The default is FALSE.

fixed

(logical) Use exact matching. The default is FALSE.

...

Additional slots and values passed to struct_class.

Value

A filter_venn object with the following output slots:

filtered	(annotation_source) Annotation_source after filtering.

flags	(data.frame) A list of flags indicating which annotations were removed.

Inheritance

A filter_venn object inherits the following struct classes:

[filter_venn] -> [model] -> [struct_class]

Examples

M <- filter_venn(
        factor_name = "V1",
        group_column = NULL,
        tables = NULL,
        levels = "",
        mode = "exclude",
        perl = FALSE,
        fixed = FALSE)