Convert to or from kegg identifiers

Searches the Kegg database to obtain external identifiers. KEGG compound, drug and glycan databases can be queried for pubchem and chebi identifiers, and vice-versa.

Usage

kegg_lookup(
  get = "pubchem",
  from = "compound",
  query_column,
  suffix = "_kegg",
  ...
)

Arguments

get

(character) Get identifier. Allowed values are limited to the following:

"compound": KEGG small molecule database.

"glycan": KEGG glycan database.

"drug": KEGG drug database.

"chebi": Chemical Entities of Biological Interest (ChEBI) database.

"pubchem": PubChem Substance Identifier.

The default is "pubchem".

from

(character) From identifier. Allowed values are limited to the following:

"compound": KEGG small molecule database.

"glycan": KEGG glycan database.

"drug": KEGG drug database.

"chebi": Chemical Entities of Biological Interest (ChEBI) database.

"pubchem": PubChem Substance Identifier.

The default is "compound".

query_column

(character) The name of the column containing identifiers to search the database for. They should be identifiers of the type selected for the "from" slot.

suffix

(character) A suffix appended to all column names in the returned result. The default is "_kegg".

...

Additional slots and values passed to struct_class.

Value

A kegg_lookup object with the following output slots:

updated	(annotation_source) An annotation_source object with a new column of compound identifiers.

Details

This object makes use of functionality from the following packages:

KEGGREST

dplyr

Inheritance

A kegg_lookup object inherits the following struct classes:

[kegg_lookup] -> [model] -> [struct_class]

References

Tenenbaum D, Maintainer B (2023). KEGGREST: Client-side REST access to the Kyoto Encyclopedia of Genes and Genomes (KEGG). doi:10.18129/B9.bioc.KEGGREST https://doi.org/10.18129/B9.bioc.KEGGREST, R package version 1.42.0, https://bioconductor.org/packages/KEGGREST.

Wickham H, François R, Henry L, Müller K, Vaughan D (2023). dplyr: A Grammar of Data Manipulation. R package version 1.1.4, https://CRAN.R-project.org/package=dplyr.

See also

Examples

M <- kegg_lookup(
        get = "pubchem",
        from = "compound",
        query_column = "V1",
        suffix = "_kegg")