Select columns

A wrapper around tidyselect::eval_select. Remove columns from an annotation table using tidy grammar.

Usage

remove_columns(expression = everything(), ...)

Arguments

expression

(call) A valid rlang::expr for tidy evaluation via eval_select. e.g. expression = all_of(c("foo","bar")) will select columns named "foo" and "bar" from the annotation data.frame. . The default is everything().

...

Additional slots and values passed to struct_class.

Value

A remove_columns object with the following output slots:

updated	(annotation_source) The updated annotations as an annotation_source object.

Details

This object makes use of functionality from the following packages:

tidyselect

rlang

Inheritance

A remove_columns object inherits the following struct classes:

[remove_columns] -> [model] -> [struct_class]

References

Henry L, Wickham H (2024). tidyselect: Select from a Set of Strings. R package version 1.2.1, https://CRAN.R-project.org/package=tidyselect.

Henry L, Wickham H (2024). rlang: Functions for Base Types and Core R and 'Tidyverse' Features. R package version 1.1.3, https://CRAN.R-project.org/package=rlang.

See also

Examples

M <- remove_columns(
        expression = call("example"))